+++ Core/Src/tm1637.c
@@ -11,7 +11,7 @@ static void terminate(void);
 int TM1637_DisplayErr();
 int TM1637_DisplayDigits(uint8_t first_digit, uint8_t second_digit, uint8_t third_digit, uint8_t fourth_digit, uint8_t show_colon);
 
-int TM1637_SetDisplay(uint32_t on) {
+uint8_t TM1637_SetDisplay(uint32_t on) {
 	/* 10001000 or 0x88 (Display On) */
 	start();
 	if (on) {
